stt/requirements.txt
openai == 1.3.5
grpcio == 1.60.0
google == 3.0.0
protobuf == 4.25.1
celery[redis] == 5.3.6
flower == 2.0.1
python-json-logger == 2.0.7
eventlet == 0.34.2